libathemecore: recvq/sendq functions should use size_t for lengths

+++ include/datastream.h
@@ -8,7 +8,7 @@
 #ifndef __CLARODATASTREAM
 #define __CLARODATASTREAM
 
-E void sendq_add(connection_t *cptr, char *buf, int len);
+E void sendq_add(connection_t *cptr, char *buf, size_t len);
 E void sendq_add_eof(connection_t *cptr);
 E void sendq_flush(connection_t *cptr);
 E bool sendq_nonempty(connection_t *cptr);
@@ -16,8 +16,8 @@ E void sendq_set_limit(connection_t *cptr, size_t len);
 
 E int recvq_length(connection_t *cptr);
 E void recvq_put(connection_t *cptr);
-E int recvq_get(connection_t *cptr, char *buf, int len);
-E int recvq_getline(connection_t *cptr, char *buf, int len);
+E int recvq_get(connection_t *cptr, char *buf, size_t len);
+E int recvq_getline(connection_t *cptr, char *buf, size_t len);
 
 E void sendqrecvq_free(connection_t *cptr);
